+++ main.rkt
@@ -12,6 +12,7 @@
          python-code
          racket-code
          rhombus-code
+         swift-code
          wasm-code
          shell-code
          scribble-code
@@ -27,6 +28,7 @@
          pythonblock
          racketblock
          rhombusblock
+         swiftblock
          wasmblock
          shellblock
          scribbleblock
@@ -42,6 +44,7 @@
          pythonblock0
          racketblock0
          rhombusblock0
+         swiftblock0
          wasmblock0
          shellblock0
          scribbleblock0
@@ -68,6 +71,7 @@
 (define-syntax-rule (python-code . rest) (api:python-code . rest))
 (define-syntax-rule (racket-code . rest) (api:racket-code . rest))
 (define-syntax-rule (rhombus-code . rest) (api:rhombus-code . rest))
+(define-syntax-rule (swift-code . rest) (api:swift-code . rest))
 (define-syntax-rule (wasm-code . rest) (api:wasm-code . rest))
 (define-syntax-rule (shell-code . rest) (api:shell-code . rest))
 (define-syntax-rule (scribble-code . rest) (api:scribble-code . rest))
@@ -83,6 +87,7 @@
 (define-syntax-rule (pythonblock . rest) (api:pythonblock . rest))
 (define-syntax-rule (racketblock . rest) (api:racketblock . rest))
 (define-syntax-rule (rhombusblock . rest) (api:rhombusblock . rest))
+(define-syntax-rule (swiftblock . rest) (api:swiftblock . rest))
 (define-syntax-rule (wasmblock . rest) (api:wasmblock . rest))
 (define-syntax-rule (shellblock . rest) (api:shellblock . rest))
 (define-syntax-rule (scribbleblock . rest) (api:scribbleblock . rest))
@@ -98,6 +103,7 @@
 (define-syntax-rule (pythonblock0 . rest) (api:pythonblock0 . rest))
 (define-syntax-rule (racketblock0 . rest) (api:racketblock0 . rest))
 (define-syntax-rule (rhombusblock0 . rest) (api:rhombusblock0 . rest))
+(define-syntax-rule (swiftblock0 . rest) (api:swiftblock0 . rest))
 (define-syntax-rule (wasmblock0 . rest) (api:wasmblock0 . rest))
 (define-syntax-rule (shellblock0 . rest) (api:shellblock0 . rest))
 (define-syntax-rule (scribbleblock0 . rest) (api:scribbleblock0 . rest))